+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+* Add labelType to PrintData
+* Add image to example desing image of canvas
+
 ## 0.0.1
 
 * Upload all files